refactor: build logic
Splits intermediate type building logic to its own file.
