+++ .gitignore
@@ -4,3 +4,4 @@ results/
 .history
 *.so
 *.bc
+tags
